Rewrite ProjectCarrousel with new shared ScrollableCarrousel component
